+++ pyacal/_control_systems/simulation/__init__.py
@@ -1,8 +1,9 @@
 """Epics-related subpackage."""
 
-from threading import Thread
 import multiprocessing as _mp
+from threading import Thread
 
+from .._handle_pvs import create_pv, get_pv
 from .pv import PV
 
 del pv
